+++ lib/src/commands/add.dart
@@ -5,7 +5,6 @@
 // found in the LICENSE file in the root of this package.
 
 import 'dart:io';
-
 import 'package:args/command_runner.dart';
 import 'package:gg_console_colors/gg_console_colors.dart';
 import 'package:gg_log/gg_log.dart';